solve invite page build errors
